// src/@types/enums/endPoints.ts
export enum EndPointsEnums {
  BANK = "Bank",
  BANK_CARD = "BankCard",
  BRANCH = "Branch",
  COUNTRY = "Country",
  All_COUNTRY = "Country/GetAll",
  CUSTOMER = "Customer",
  CUSTOMER_CODE = "Customer/GetNewCode",
  ALL_CUSTOMERS = "Customer/GetAll",
  CUSTOMER_GROUP = "CustomerGroup",
  ALL_CUSTOMER_GROUPs = "CustomerGroup/GetAll",
  EMPLOYEE = "Employee",
  EXPENSE = "Expense",
  REPRESENTATIVE = "SalesMan",
  ALL_REPRESENTATIVE = "SalesMan/GetAll",
  JOURNAL = "Journal",
  STORE_KEEPER = "StoreKeeper",
  ALL_STORE_KEEPERS = "StoreKeeper/GetAll",
  RoleUser = "RoleUser",
  AllRoleUser = "RoleUser/GetAll",
  AllAccount = "Account/GetAll",
  ALLPAYMENT = "Payment/GetAll",
  PAYMENT_CODE = "Payment/GetNewCode",
  TAX_TYPE = "TaxType",
  ALL_TAX_TYPE = "TaxType/GetAll",
  User = "User",
  AllUser = "User/GetAll",
  AllBank = "Bank/GetAll",
  AllTax = "Tax/GetAll",
  AllStore = "Store/GetAll",
  AllBranch = "Branch/GetAll",
  AllBankCards = "BankCard/GetAll",
  Store = "Store",
  ALL_Expense = "Expense/GetAll",
  VENDOR = "Vendor",
  VENDOR_CODE = "Vendor/GetNewCode",
  ALL_VENDORS = "Vendor/GetAll",
  SAFE = "Safe",
  ALLSafe = "Safe/GetAll",
  ACCOUNT = "Account",
  VENDOR_GROUP = "VendorGroup",
  ALL_VENDOR_GROUPS = "VendorGroup/GetAll",
  LIST_VALUE_VENDOR_PRICE = "ListValueVendorPrices",
  ALL_LIST_VALUE_VENDOR_PRICES = "ListValueVendorPrices/GetAll",
  LIST_VALUE_CUSTOMER_PRICE = "ListValueCustomerPrices",
  ALL_LIST_VALUE_CUSTOMER_PRICES = "ListValueCustomerPrices/GetAll",
  LIST_VALUE_PAYMENT_METHOD = "ListValuePaymentMethods",
  ALL_LIST_VALUE_PAYMENT_METHODS = "ListValuePaymentMethods/GetAll",
  ALL_PAYMENT_METHODS = "PaymentMethods/GetAll",
  LIST_INVOICE_DEAL_TYPE = "ListInvoiceDealTypes",
  ALL_LIST_INVOICE_DEAL_TYPE = "ListInvoiceDealTypes/GetAll",
  LIST_VALUE_COMMISSION_TYPE = "ListCommissionTypes",
  ALL_LIST_VALUE_COMMISSION_TYPE = "ListCommissionTypes/GetAll",
  LIST_VALUE_TYPES = "ListValueTypes",
  ALL_LIST_VALUE_TYPES = "ListValueTypes/GetAll",
  REGION = "Region",
  ALL_REGION = "Region/GetAll",
  UNIT_TYPE = "UnitItemType",
  ALL_UNIT_TYPE = "UnitItemType/GetAll",
  PRODUCT = "Product",
  ALL_PRODUCTS = "Product/GetAll",
  PRODUCT_GROUP = "ProductGroup",
  ALL_PRODUCT_GROUP = "ProductGroup/GetAll",
  ALL_PURCHASE_INVOICE = "PurchaseInvoice/GetAll",
  PURCHASE_INVOICE = "PurchaseInvoice",
  PURCHASE_INVOICE_CODE = "PurchaseInvoice/GetNewCode",
  ALL_PURCHASE_RETURN_INVOICE = "PurchaseReturnInvoice/GetAll",
  PURCHASE_RETURN_INVOICE = "PurchaseReturnInvoice",
  PURCHASE_RETURN_INVOICE_CODE = "PurchaseReturnInvoice/GetNewCode",
  PURCHASE_INVOICE_REPORT = "PurchaseReport/PurchaseInvoiceReport",
  PURCHASE_RETUTN_INVOICE_REPORT = "PurchaseReturnReport/PurchaseReturnInvoiceReport",
  PURCHASE_ORDER = "PurchesesOrder",
  PURCHASE_ORDER_CODE = "PurchesesOrder/GetNewCode",
  OUTGOING_STORE_TRANSACTION = "OutgoingStoreTransaction",
  TRANSFER_STORE_TRANSACTION = "TransferStoreTransaction",
  INCOMING_STORE_TRANSACTION = "IncomingStoreTransaction",
  LIST_VALUE_OUTGOING_STORE_TYPES = "ListValueOutgoingStoreTypes",
  LIST_VALUE_INCOMING_STORE_TYPES = "ListValueIncomingStoreTypes/GetAll",
  PRODUCT_UNIT = "ProductUnit",
  // SALES_MAN = 'SalesMan',
  SALES_INVOICE = "SalesInvoice",
  ALL_SALES_INVOICES = "SalesInvoice/GetAll",
  SALES_INVOICE_CODE = "SalesInvoice/GetNewCode",
  ALL_SALES_RETURN_INVOICES = "SalesReturnInvoice/GetAll",
  SALES_RETURN_INVOICE = "SalesReturnInvoice",
  SALES_RETURN_INVOICE_CODE = "SalesReturnInvoice/GetNewCode",
  SALES_ORDER = "OfferPrice",
  SALES_ORDER_CODE = "OfferPrice/GetNewCode",
  PAYMENT = "Payment",
  RECEIPT_VOUCHER = "ReceiptVoucher",
  ALL_RECEIPT_VOUCHER = "ReceiptVoucher/GetAll",
  RECEIPT_VOUCHER_CODE = "ReceiptVoucher/GetNewCode",
  ALL_INCOMING_STORE_TRANSACTION = "IncomingStoreTransaction/GetAll",
  INCOMING_STORE_TRANSACTION_CODE = "IncomingStoreTransaction/GetNewCode",
  PRODUCT_CODE = "Product/GetNewCode",
  ALL_OUTGOING_STORE_TRANSACTION = "OutgoingStoreTransaction/GetAll",
  OUTGOING_STORE_TRANSACTION_CODE = "OutgoingStoreTransaction/GetNewCode",
  ALL_STORE_TRANSFER_TRANSACTION = "TransferStoreTransaction/GetAll",
  STORE_TRANSFER_TRANSACTION_CODE = "TransferStoreTransaction/GetNewCode",
}